requirements.txt
plotly==5.10.0
streamlit-aggrid==0.3.3
openpyxl